New: First step in re-writting the CantusEditor main html/javascript
Major changes to the index.html file.

Removing all `hardcoded` paths. Replacing them for three js objects
that store the essential information about manuscripts, mei sources,
and IIIF sources. In the next steps, the js, html, and css will be
separated into different files and the code refactored for the
remaining sections of the javascript code.